chore(storage): rename createStore to create
