+++ packages/storage/src/storage-browser/index.ts
@@ -7,7 +7,7 @@ export {
 	ListCallerAccessGrantsOutput,
 } from './listCallerAccessGrants';
 export { createLocationCredentialsHandler } from './createLocationCredentialsHandler';
-export { createLocationCredentialsStore } from './locationCredentialsStore/createStore';
+export { createLocationCredentialsStore } from './locationCredentialsStore/create';
 export {
 	managedAuthAdapter,
 	ManagedAuthAdapterInput,
